Find projector/mirror screen based on model name
For things to work correctly with xrandr, we need to *first* configure the
mirror screen (2240x1400), *then* the projector screen (usually full HD or
1920x1200). Otherwise, there is weird overlap between mirror and internal
screen.

However, that means we need to find out which is which if multiple USB-C screens
are connected. Thanks to xrandr showing the model name and jc extracting it,
things got much easier.